toyos-ld: Mach-O TLS & SUBTRACTOR support, archive input, ELF TLS fix
- Add Mach-O SUBTRACTOR relocation pairs (difference expressions)
- Add Mach-O TLS sections: __thread_vars, __thread_data, __thread_bss
- Classify ARM64 PAGEOFF12 by inspecting instruction (ADD vs LDR/STR)
- Add ARM64_RELOC_POINTER_TO_GOT and TLVP_LOAD_PAGE21 relocation types
- Flatten static archives (.a) into individual objects before linking
- Fix ELF TLS layout: place .tdata within RW segment for correct PT_TLS
- Refactor MachOLayout to use Vec<MachOSection> instead of individual fields
- Handle Common section kind, explicit section kind matching with panic on unknown
- Remove outdated integration tests

+++ userland/toyos-ld/src/emit_pe.rs
@@ -194,7 +194,7 @@ pub(crate) fn emit_pe_bytes(
         for (sym_ref, &got_vaddr) in &layout.got {
             let sym_addr = resolve_symbol(state, sym_ref, None)
                 .ok_or_else(|| LinkError::UndefinedSymbols(vec![sym_ref.name().to_string()]))?;
-            let off = (got_vaddr as u32 - layout.data_rva) as usize;
+            let off = (got_vaddr - layout.data_rva as u64) as usize;
             data_data[off..off + 8].copy_from_slice(&sym_addr.to_le_bytes());
         }
         w.write_section(data_range.file_offset, &data_data);
